Running Jasmine spec through production UI vol. I
Goal here is to be able to run all test (not just on localhost but) in
production using a special parameter (probably ?test=true). When that
parameter is set then all test resources (js, css) are loaded in addition to
production resources, and all tests are run. Most of the ajax calls are
faked during the tests.

Also when someone has an issue she can check easily if the tests passes in
her environment (browser version etc.).

Thanks @eeroan for the original idea.

This is an in-between feature commit. Obviously still need to resolve on
page load which resources should be loaded. Bare with me.